requirements.txt
asgiref==3.8.1
click==8.1.8
colorama==0.4.6
coverage==7.8.2
cssbeautifier==1.15.1
diff-match-patch==20241021
dj-database-url==2.3.0
Django==5.1.5
django-cors-headers==4.6.0
django-filter==25.1
django-import-export==4.3.7
djangorestframework==3.16.0
djlint==1.36.4
EditorConfig==0.17.0
graphene==3.4.3
graphene-django==3.2.2
graphql-core==3.2.5
graphql-relay==3.2.0
gunicorn==23.0.0
jsbeautifier==1.15.1
json5==0.10.0
Markdown==3.8
packaging==25.0
pathspec==0.12.1
promise==2.3
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-decouple==3.8
PyYAML==6.0.2
regex==2024.11.6
six==1.17.0
sqlparse==0.5.3
tablib==3.8.0
text-unidecode==1.3
tqdm==4.67.1
typing_extensions==4.12.2
